+++ torchvision/ops/__init__.py
@@ -4,6 +4,10 @@ from .roi_pool import roi_pool, RoIPool
 from .poolers import MultiScaleRoIAlign
 from .feature_pyramid_network import FeaturePyramidNetwork
 
+from ._register_onnx_ops import _register_custom_op
+
+_register_custom_op()
+
 
 __all__ = [
     'nms', 'roi_align', 'RoIAlign', 'roi_pool', 'RoIPool',
